feat(search): live grep over playground
You can now use `:PlaygroundSearch` to do a live grep over the files
inside of the playgrounds. There will need to be more steps in the
future if I am to make sure you can still do the build steps. Will try
to figure out a good compromise for this.